requirements.txt
fastapi[standard]
python-multipart
pymupdf4llm
pymupdf_layout
easyocr
langchain-openai
psycopg[binary]
langchain-text-splitters
pydantic-settings
python-dotenv
slowapi
redis
cryptography
structlog
sqlmodel
pgvector
pyjwt
passlib[argon2]
alembic
streamlit
requests
